backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Pydantic — 2.8+ required by langchain-core 1.x and pydantic-settings 2.10+
pydantic>=2.8.0
pydantic-settings>=2.10.0

# OpenAI SDK — 1.50+ for stable images.generate interface
openai>=1.50.0

# LangChain ecosystem — keep the entire stack on matching 1.x releases.
# DO NOT pin these to 0.x — they are Python-3.12-incompatible.
langchain>=1.0.0
langchain-openai>=1.0.0
langchain-core>=1.0.0
langsmith>=0.2.0        # 0.1.x has a ForwardRef/Python-3.12 crash
langgraph>=1.0.0        # 0.1.x API is incompatible with 1.x checkpoint packages

# Image processing
Pillow>=10.3.0

# Async helpers
anyio>=4.4.0

# Dev / testing
pytest>=8.0.0
pytest-asyncio>=0.23.7
httpx>=0.27.0
python-dotenv>=1.0.1
